### 2020-09-01 #2628 @whedon
[PRE REVIEW]: Xgcm: A Python Package for Analyzing Data from General Circulation Models
**Submitting author:** @rabernat (<a href="http://orcid.org/0000-0001-5999-4917">Ryan Abernathey</a>)
**Repository:** <a href="https://github.com/xgcm/xgcm" target ="_blank">https://github.com/xgcm/xgcm</a>
**Version:** v.0.3.0
