### 2020-09-01 #2628 @kthyng
Hi @rabernat! It's great to see `xgcm` here! I'll be your editor.

Can you start by checking out the doi's noted above as possible problems, and then suggest reviewers either from your own network and/or from the [list of volunteers](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit?usp=sharing). The best reviewers would be present or future users of xgcm who have not been involved with the development.
