+++ paper.bib
@@ -5,7 +5,8 @@
   volume={5},
   number={1},
   year={2017},
-  publisher={Ubiquity Press}
+  publisher={Ubiquity Press},
+  doi={10.5334/jors.148}
 }
 
 @inproceedings{rocklin2015dask,
@@ -14,7 +15,8 @@
   booktitle={Proceedings of the 14th python in science conference},
   number={130-136},
   year={2015},
-  organization={Citeseer}
+  organization={Citeseer},
+  doi={10.25080/majora-7b98e3ed-013}
 }
 
 @article{busecke_euc_omz_2019,
@@ -68,7 +70,7 @@ pages = "173 - 265",
 year = "1977",
 booktitle = "General Circulation Models of the Atmosphere",
 issn = "0076-6860",
-doi = "https://doi.org/10.1016/B978-0-12-460817-7.50009-4",
+doi = "10.1016/B978-0-12-460817-7.50009-4",
 url = "http://www.sciencedirect.com/science/article/pii/B9780124608177500094",
 author = "Akio Arakawa and Vivian R. Lamb"
 }
@@ -91,7 +93,7 @@ url = {https://github.com/ECCO-GROUP/ECCOv4-py}
 }
 
 @article{almansi_et_al_2019,
-  doi = {<a class="libx-autolink" href="http://rd8hp6du2b.search.serialssolutions.com/?url_ver=Z39.88-2004&amp;rft_val_fmt=info:ofi/fmt:kev:mtx:journal&amp;__char_set=utf8&amp;rft_id=info:doi/10.21105/joss.01506&amp;rfr_id=info:sid/libx&amp;rft.genre=article">10.21105/joss.01506</a>},
+  doi = {10.21105/joss.01506},
   url = {https://doi.org/<a class="libx-autolink" href="http://rd8hp6du2b.search.serialssolutions.com/?url_ver=Z39.88-2004&amp;rft_val_fmt=info:ofi/fmt:kev:mtx:journal&amp;__char_set=utf8&amp;rft_id=info:doi/10.21105/joss.01506&amp;rfr_id=info:sid/libx&amp;rft.genre=article">10.21105/joss.01506</a>},
   year = {2019},
   publisher = {The Open Journal},
