### 2020-09-01 #2628 @rabernat
Thanks @kthyng!

> Can you start by checking out the doi's noted above as possible problems

I have fixed the DOIs on the master branch.

> The best reviewers would be present or future users of xgcm who have not been involved with the development.

Based on your guidance and the spreadsheet, I would suggest

- @sadielbartholomew
- @edoddridge
- @koldunovn
- @wingtorres

Additional possible reviewers who are not on the JOSS list could be

- @vrx-
- @emmomp
