### 2020-09-01 5f01984
Adapted ci.yml from cf-xarray and renamed ci envs
