+++ .travis.yml
@@ -1,49 +1,0 @@
-# Based on http://conda.pydata.org/docs/travis.html
-language: python
-sudo: false # use container based build
-notifications:
-  email: false
-
-matrix:
-  fast_finish: true
-  include:
-  - python: 3.6
-    env: CONDA_ENV=py36
-  - python: 3.7
-    env: CONDA_ENV=py37
-  - python: 3.7
-    env: CONDA_ENV=py37-upstream-master
-
-before_install:
-  - wget http://repo.continuum.io/miniconda/Miniconda3-3.16.0-Linux-x86_64.sh -O miniconda.sh;
-  - bash miniconda.sh -b -p $HOME/miniconda
-  - export PATH="$HOME/miniconda/bin:$PATH"
-  - hash -r
-  - conda config --set always_yes yes --set changeps1 no
-  - conda update -q conda
-  - conda info -a
-  - conda list
-
-install:
-  - conda env create --file ci/environment-$CONDA_ENV.yml
-  - source activate test_env_xgcm
-  - conda list
-  - python -c "import xarray; print(xarray.__version__)"
-
-# if we do this, the package gets installed into
-# /home/travis/build/xgcm/xgcm/build/lib/xgcm
-#  - python setup.py install
-  - pip install -e .
-# this puts the package into
-# /home/travis/build/xgcm/xgcm/xgcm
-# That turns out to be necessary for py.test to correctly collect the tests
-# and for coverage to work properly.
-# It is very complicated and confusing.
-
-script:
-  # only check code for py36 env
-  - black --check xgcm
-  - py.test xgcm -v --cov=xgcm --cov-config .coveragerc --cov-report term-missing
-
-after_success:
-  - codecov
