### 2020-09-01 32087e2
Some style nits in the ci and linting yaml
