+++ setup.cfg
@@ -8,3 +8,16 @@ versionfile_source = xgcm/_version.py
 versionfile_build = xgcm/_version.py
 tag_prefix = v
 parentdir_prefix = xgcm-
+
+[flake8]
+exclude = __init__.py,versioneer.py,.eggs,doc
+ignore =
+    # whitespace before ':' - doesn't work well with black
+    E203
+    E402
+    # line too long - let black worry about that
+    E501
+    # do not assign a lambda expression, use a def
+    E731
+    # line break before binary operator
+    W503
