### 2020-09-01 893693d
Fix leftover typo and minor style
