+++ .github/workflows/linting.yaml
@@ -22,7 +22,7 @@ jobs:
           python-version: 3.8
       - name: Install Black
         run: |
-          python -m pip install black<=
+          python -m pip install black
       - name: Black Code Formatter
         run: |
           black --check . -v
