### 2020-09-01 657bfc2
Deactivate flake8 in favor of sep PR
