### 2020-09-01 95d4518
Updated pre-commit and linting according to suggestions
