### 2020-09-01 #2628 @kthyng
Sorry for the github pings everyone! (we list github handles without the "@" to not email people at this stage to limit noise for potential reviewers, but I forgot to say that)

I am going to narrow down @rabernat's list, especially to avoid people who are already actively reviewing for us!

@koldunovn and @vrx- would you be interested in reviewing this package? The review process is all on github and is focused on the code itself. We're asking for reviews within 5-6 weeks (longer than usual due to the on-going pandemic). You can find more information [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

@vrx- I know you have already used xgcm so I'm hoping you can go into depth in particular in your feedback.
@koldunovn I hope that you haven't used xgcm before but do use ocean models since then you would be a good sample future user for such a package.
