### 2020-09-01 80bae86
Added new env files to ci for other python versions
