+++ ci/environment-py3.8.yml
@@ -1,0 +1,15 @@
+name: test_env_xgcm
+dependencies:
+  - python=3.8
+  - xarray
+  - dask
+  - numpy
+  - pytest
+  - future
+  - pip
+  - black
+  - pip:
+    - docrep<=0.2.7
+    - codecov
+    - pytest-cov
+
