+++ ci/environment-upstream-dev.yml
@@ -1,6 +1,6 @@
 name: test_env_xgcm
 dependencies:
-  - python=3.7
+  - python
   - dask
   - numpy
   - pytest
