### 2020-09-01 1f5e3c8
Deactivate 3.9 for now
