### 2020-09-01 ad30284
Remove 3.9. Its not actually released
