+++ ci/environment-py3.9.yml
@@ -1,15 +1,0 @@
-name: test_env_xgcm
-dependencies:
-  - python=3.9
-  - xarray
-  - dask
-  - numpy
-  - pytest
-  - future
-  - pip
-  - black
-  - pip:
-    - docrep<=0.2.7
-    - codecov
-    - pytest-cov
-
