+++ .github/workflows/ci.yaml
@@ -17,7 +17,7 @@ jobs:
       fail-fast: false
       matrix:
         os: ["ubuntu-latest"]
-        python-version: ["3.6", "3.7", "3.8"] # 3.9 causes problems for now
+        python-version: ["3.6", "3.7", "3.8"] # 3.9 is not yet released
     steps:
       - uses: actions/checkout@v2
       - name: Cache conda
