### 2020-09-01 ecc816a
Modernized setup.py
