### 2020-09-01 537d347
Fix typo in setup.py
