### 2020-09-01 68dd1f1
Codecov should not fail the whole build
