### 2020-09-01 ef3497f
Move codecov options to seperate file
