+++ .github/workflows/ci.yaml
@@ -9,23 +9,6 @@ on:
   schedule:
     - cron: "0 13 * * 1"
 
-codecov:
-  require_ci_to_pass: no
-  max_report_age: off
-
-comment: false
-
-coverage:
-  precision: 2
-  round: down
-  status:
-    project:
-      default:
-        target: 95
-        informational: true
-    patch: off
-    changes: off
-
 jobs:
   build:
     name: Build (${{ matrix.python-version }} | ${{ matrix.os }})
