### 2020-09-02 553e84d
Migrate CI to github actions (#214)
* Pin docrep < 0.3.0

* Pinned old docrep v in ci envs

* List conda versions in ci after env install

* Added explicit pip and switched black to conda dependencies

* Forgot upstream ci env

* Adapted ci.yml from cf-xarray and renamed ci envs

* Some style nits in the ci and linting yaml

* Tame flake 8

* Focus on black for now

* The new black likes one line docstrings differently

* Reactivate flake 8

* Fix leftover typo and minor style

* Deactivate flake8 in favor of sep PR

* Updated pre-commit and linting according to suggestions

* Added new env fi
[…]
