+++ ci/environment-py3.7.yml
@@ -1,13 +1,15 @@
 name: test_env_xgcm
 dependencies:
   - python=3.7
+  - xarray
   - dask
   - numpy
   - pytest
   - future
+  - pip
+  - black
   - pip:
-    - docrep
+    - docrep<=0.2.7
     - codecov
     - pytest-cov
-    - black
-    - git+https://github.com/pydata/xarray.git
+
