+++ .pre-commit-config.yaml
@@ -1,6 +1,5 @@
 repos:
-  - repo: https://github.com/ambv/black
+  - repo: https://github.com/python/black
     rev: stable
     hooks:
     - id: black
-      language_version: python3.8
