+++ ci/environment-py3.6.yml
@@ -1,13 +1,15 @@
 name: test_env_xgcm
 dependencies:
-  - python=3.7
+  - python=3.6
   - xarray
   - dask
   - numpy
   - pytest
   - future
+  - pip
+  - black
   - pip:
-    - docrep
+    - docrep<=0.2.7
     - codecov
     - pytest-cov
-    - black
+
