+++ doc/environment.yml
@@ -15,7 +15,7 @@ dependencies:
   - sphinx_rtd_theme
   - ipykernel
   - pip:
-    - docrep
+    - docrep<=0.2.7
     - nbsphinx
     - jupyter_client
     - sphinx_pangeo_theme

[…]
