### 2020-09-02 #2631 @kthyng
Here we are in the review @koldunovn, @vrx-! You can see your checklists above to drive the review process. As you find issues to raise as part of the review, please share them by starting a github issue in the software repo itself, then link it back to this review issue by including a link to here. This will help keep everything organized. Then you can write a small summary of the issues you've raised in the repo. Let me know if you have questions, esp since this is a unique reviewing approach!
