+++ paper.bib
@@ -70,11 +70,7 @@ pages = "173 - 265",
 year = "1977",
 booktitle = "General Circulation Models of the Atmosphere",
 issn = "0076-6860",
-<<<<<<< HEAD
-doi = "https://doi.org/10.1016/B978-0-12-460817-7.50009-4",
-=======
 doi = "10.1016/B978-0-12-460817-7.50009-4",
->>>>>>> joss-paper
 url = "http://www.sciencedirect.com/science/article/pii/B9780124608177500094",
 author = "Akio Arakawa and Vivian R. Lamb"
 }
