### 2020-09-02 #2631 @rabernat
I introduced a typo into our bib file when addressing https://github.com/openjournals/joss-reviews/issues/2628#issuecomment-685008307. This has been fixed in https://github.com/xgcm/xgcm/commit/5634d71af4f45a86062d6b0bbb2138ea416b112a. I have verified that the preview now builds using  https://whedon.theoj.org/.
