### 2020-09-02 63ea692
Update whats-new.rst
