### 2020-09-03 62bf857
new conda forge badge (rounded corners)
