### 2020-09-08 e046e2a
add numba to ci
