+++ ci/environment-py38.yml
@@ -1,14 +1,0 @@
-name: test_env_xgcm
-dependencies:
-  - python=3.8
-  - xarray
-  - dask
-  - numpy
-  - numba
-  - pytest
-  - future
-  - pip:
-    - docrep
-    - codecov
-    - pytest-cov
-    - black
