### 2020-09-08 44cfbbc
add numba to dev env
