+++ ci/environment-py38-upstream-master.yml
@@ -1,15 +1,0 @@
-name: test_env_xgcm
-dependencies:
-  - python=3.8
-  - xarray
-  - dask
-  - numpy
-  - numba
-  - pytest
-  - future
-  - pip
-  - black
-  - pip:
-    - docrep<=0.2.7
-    - codecov
-    - pytest-cov
