+++ xgcm/test/test_vertical_grid.py
@@ -1,49 +1,0 @@
-"""
-Tests for vertical grid routines.
-"""
-
-import pytest
-
-import numpy as np
-from ..vertical_grid import interp_1d_linear, interp_1d_conservative
-
-
-def test_interp_1d_linear():
-    nz, nx = 100, 1000
-    z_vertex = np.linspace(0, 1, nz + 1)
-    z = 0.5 * (z_vertex[:-1] + z_vertex[1:])
-    x = 2 * np.pi * np.linspace(0, 1, nx)
-    # uniformly stratified scalar
-    theta = z + 0.1 * np.cos(3 * x)[:, None]
-    # the scalar to interpolate
-    phi = np.sin(theta) + 0.1 * np.cos(5 * x)[:, None]
-    target_theta_levels = np.arange(0.2, 0.9, 0.025)
-    phi_at_theta_expected = np.sin(target_theta_levels) + 0.1 * np.cos(5 * x)[:, None]
-
-    # the action
-    phi_at_theta = interp_1d_linear(
-        phi, theta, target_theta_levels, mask_outcrops=False
-    )
-    np.testing.assert_allclose(phi_at_theta, phi_at_theta_expected, rtol=1e-4)
-
-
-def test_interp_1d_conservative():
-    nz = 30
-    k = np.arange(nz)
-    dz = 10 + np.linspace(0, 90, nz - 1)
-    z = np.concatenate([[0], np.cumsum(dz)])
-    H = z.max()
-    theta = z / H + 0.2 * np.cos(np.pi * z / H)
-    # phi = np.sin(5 * np.pi * z/H)
-
-    nbins = 100
-    theta_bins = np.linspace(theta.min() - 0.1, theta.max() + 0.1, nbins)
-
-    # lazy way to check that it vectorizes: just copy the 1d array
-    nx = 5
-    dz_2d = np.tile(dz, (nx, 1))
-    theta_2d = np.tile(theta, (nx, 1))
-
-    dz_theta = interp_1d_conservative(dz_2d, theta_2d, theta_bins)
-
-    np.testing.assert_allclose(dz_theta.sum(axis=-1), dz.sum(axis=-1))
