### 2020-09-08 9e59cbb
Added grid method
