### 2020-09-08 72ea196
move input handling to decorator
