### 2020-09-08 6b63c0e
Added first hacky test for conservative interpolation
