+++ xgcm/test/test_transform.py
@@ -87,9 +87,6 @@ def raw_datasets():
     # Second set of examples: Conservative remapping to other depth levels.
     data = np.array([1, 4, 0])
     target_z_bounds = np.array([0, 1, 10, 50, 80])
-    target_z = (
-        target_z_bounds[1:] + target_z_bounds[0:-1]
-    ) / 2  # simply infer the cell center
     input_conservative_depth_depth = xr.Dataset(
         {"data": xr.DataArray(data, dims=["z"], coords={"z": z})}
     )
@@ -102,7 +99,9 @@ def raw_datasets():
     expected_conservative_depth_depth = xr.DataArray(
         np.array([0.1, 0.9, 4.0, 0.0]),
         dims=["zc"],
-        coords={"zc": target_z},
+        coords={
+            "zc": (target_z_bounds[1:] + target_z_bounds[0:-1]) / 2
+        },  # simply infer the cell center
     )
 
     datasets = {
@@ -110,28 +109,28 @@ def raw_datasets():
             input_linear_depth_depth,
             {"coords": {"Z": {"center": "z"}}},
             target_z,
-            {"mask_edges": False},
+            {"mask_edges": False, "method": "linear"},
             expected_linear_depth_depth,
         ),
         "linear_depth_depth_renamed": (
             input_linear_depth_depth.rename({"z": "test"}),
             {"coords": {"Z": {"center": "test"}}},
             target_z,
-            {"mask_edges": True},
+            {"mask_edges": True, "method": "linear"},
             expected_linear_depth_depth_masked_outcrops.rename({"z": "test"}),
         ),
         "linear_depth_depth_masked_outcrops": (
             input_linear_depth_depth,
             {"coords": {"Z": {"center": "z"}}},
             target_z,
-            {"mask_edges": True},
+            {"mask_edges": True, "method": "linear"},
             expected_linear_depth_depth_masked_outcrops,
         ),
         "conservative_depth_depth": (
             input_conservative_depth_depth,
             {"coords": {"Z": {"inner": "z", "outer": "zc"}}},
             target_z_bounds,
-            {},
+            {"method": "conservative"},
             expected_conservative_depth_depth,
         ),
     }
@@ -144,6 +143,7 @@ def raw_datasets():
         "linear_depth_depth",
         "linear_depth_depth_masked_outcrops",
         "linear_depth_depth_renamed",
+        "conservative_depth_depth",
     ],
 )
 def dataset(request):
@@ -161,6 +161,8 @@ def test_linear_interpolation_target_value_error():
     input, grid_kwargs, target, transform_kwargs, expected = raw_datasets()[
         "linear_depth_depth"
     ]
+    # method keyword is only for high level tests
+    transform_kwargs.pop("method")
     with pytest.raises(ValueError):
         interpolated = linear_interpolation(input.data, input["z"], target, "z", "z")
     # TODO: test with the other method
@@ -179,7 +181,10 @@ def test_low_level_linear(name):
     input, grid_kwargs, target, transform_kwargs, expected = raw_datasets()[name]
     dim = grid_kwargs["coords"]["Z"]["center"]
     da_target = xr.DataArray(target, dims=[dim], coords={dim: target})
-    print(transform_kwargs)
+
+    # method keyword is only for high level tests
+    transform_kwargs.pop("method")
+
     interpolated = linear_interpolation(
         input.data, input[dim], da_target, dim, dim, dim, **transform_kwargs
     )
@@ -198,6 +203,10 @@ def test_low_level_conservative(name):
     dim = grid_kwargs["coords"]["Z"]["inner"]
     target_dim = grid_kwargs["coords"]["Z"]["outer"]
[…]
