### 2020-09-09 ad93b09
Refactored test dataset creation
