+++ xgcm/test/test_transform.py
@@ -55,102 +55,154 @@ def test_interp_1d_conservative():
     np.testing.assert_allclose(dz_theta.sum(axis=-1), dz.sum(axis=-1))
 
 
-"""1D Test datasets for various transformations."""
+"""1D Test datasets for various transformations.
+This part of the module is organized as follows:
+
+1. A nested dictionary with all parameters needed to construct a set of test datasets
+2. Fixtures that use a constructor function and all or a subset of the dictionary entries from 1. to returns:
+     (input_dataset, grid_kwargs, target, transform_kwargs, expected_dataarray)
+3. Test functions
+
+"""
+cases = {
+    "linear_depth_depth": {
+        "input_coord": ("z", [5, 25, 60]),
+        "input_data": ("data", [0.23246861, 0.45175654, 0.58320681]),  # random numbers
+        "target_coord": ("z", [0, 7, 30, 60, 70]),
+        "target_data": ("z", [0, 7, 30, 60, 70]),
+        "expected_coord": ("z", [0, 7, 30, 60, 70]),
+        "expected_data": (
+            "data",
+            np.interp(
+                [0, 7, 30, 60, 70], [5, 25, 60], [0.23246861, 0.45175654, 0.58320681]
+            ),
+        ),  # same input as in `input_data`
+        "grid_kwargs": {"coords": {"Z": {"center": "z"}}},
+        "transform_kwargs": {"mask_edges": False, "method": "linear"},
+    },
+    "linear_depth_depth_masked_outcrops": {
+        "input_coord": ("z", [5, 25, 60]),
+        "input_data": ("data", [0.23246861, 0.45175654, 0.58320681]),  # random numbers
+        "target_coord": ("z", [0, 7, 30, 60, 70]),
+        "target_data": ("z", [0, 7, 30, 60, 70]),
+        "expected_coord": ("z", [0, 7, 30, 60, 70]),
+        "expected_data": (
+            "data",
+            np.interp(
+                [0, 7, 30, 60, 70], [5, 25, 60], [0.23246861, 0.45175654, 0.58320681]
+            ),
+        ),  # same input as in `input_data`
+        "expected_data_mask_index": [0, -1],
+        "expected_data_mask_value": np.nan,
+        "grid_kwargs": {"coords": {"Z": {"center": "z"}}},
+        "transform_kwargs": {"mask_edges": True, "method": "linear"},
+    },
+    "linear_depth_depth_renamed": {
+        "input_coord": ("test", [5, 25, 60]),
+        "input_data": ("data", [0.23246861, 0.45175654, 0.58320681]),  # random numbers
+        "target_coord": ("something", [0, 7, 30, 60, 70]),
+        "target_data": ("something", [0, 7, 30, 60, 70]),
+        "expected_coord": ("something", [0, 7, 30, 60, 70]),
+        "expected_data": (
+            "data",
+            np.interp(
+                [0, 7, 30, 60, 70], [5, 25, 60], [0.23246861, 0.45175654, 0.58320681]
+            ),
+        ),  # same input as in `input_data`
+        "grid_kwargs": {"coords": {"Z": {"center": "test"}}},
+        "transform_kwargs": {"mask_edges": False, "method": "linear"},
+    },
+    "conservative_depth_depth": {
+        "input_coord": ("z", [5, 25, 60]),
+        "input_bounds_coord": ("zc", [0, 10, 50, 75]),
+        "input_data": ("data", [1, 4, 0]),
+        "target_coord": ("z", [0, 1, 10, 50, 80]),
+        "target_data": ("z", [0, 1, 10, 50, 80]),
+        "expected_coord": ("z", [0.5, 5.5, 30, 65]),
[…]
