### 2020-09-09 235d0ba
Wrapped cases dict in function to avoid modification
