+++ xgcm/test/test_transform.py
@@ -130,6 +130,67 @@ def cases():
             "grid_kwargs": {"coords": {"Z": {"center": "test"}}},
             "transform_kwargs": {"mask_edges": False, "method": "linear"},
         },
+        # example of interpolating onto a tracer that increases with depth
+        # but with inverted target
+        "linear_depth_dens": {
+            "input_coord": ("depth", [5, 25, 60, 80, 100, 120]),
+            "input_data": ("data", [1, 4, 6, 2, 0, -3]),
+            "input_additional_data_coord": ("depth", [5, 25, 60, 80, 100, 120]),
+            "input_additional_data": ("dens", [1, 5, 10, 20, 24, 35]),
+            "target_coord": ("something", [0, 5, 10, 11, 15, 20, 25, 27]),
+            "target_data": ("something", [0, 5, 10, 11, 15, 20, 25, 27]),
+            "expected_coord": ("something", [0, 5, 10, 11, 15, 20, 25, 27]),
+            "expected_data": (
+                "data",
+                [1.0, 4.0, 6.0, 5.6, 4.0, 2.0, -0.272727, -0.818182],
+            ),
+            "grid_kwargs": {
+                "coords": {"Z": {"center": "depth", "outer": "depth_bnds"}}
+            },
+            "transform_kwargs": {"method": "linear", "target_data": "dens"},
+        },
+        # example of interpolating onto a tracer that increases with depth
+        # with masked values
+        "linear_depth_dens_masked": {
+            "input_coord": ("depth", [5, 25, 60, 80, 100, 120]),
+            "input_data": ("data", [1, 4, 6, 2, 0, -3]),
+            "input_additional_data_coord": ("depth", [5, 25, 60, 80, 100, 120]),
+            "input_additional_data": ("dens", [1, 5, 10, 20, 24, 35]),
+            "target_coord": ("something", [0, 5, 10, 11, 15, 20, 25, 27]),
+            "target_data": ("something", [0, 5, 10, 11, 15, 20, 25, 27]),
+            "expected_coord": ("something", [0, 5, 10, 11, 15, 20, 25, 27]),
+            "expected_data": (
+                "data",
+                [np.nan, 4.0, 6.0, 5.6, 4.0, 2.0, -0.272727, -0.818182],
+            ),
+            "grid_kwargs": {
+                "coords": {"Z": {"center": "depth", "outer": "depth_bnds"}}
+            },
+            "transform_kwargs": {
+                "method": "linear",
+                "target_data": "dens",
+                "mask_edges": True,
+            },
+        },
+        # example of interpolating onto a tracer that increases with depth
+        # but with inverted target
+        "linear_depth_dens_reverse": {
+            "input_coord": ("depth", [5, 25, 60, 80, 100, 120]),
+            "input_data": ("data", [1, 4, 6, 2, 0, -3]),
+            "input_additional_data_coord": ("depth", [5, 25, 60, 80, 100, 120]),
+            "input_additional_data": ("dens", [1, 5, 10, 20, 24, 35]),
+            "target_coord": ("something", [27, 25, 20, 15, 11, 10, 5, 0]),
+            "target_data": ("something", [27, 25, 20, 15, 11, 10, 5, 0]),
+            "expected_coord": ("something", [27, 25, 20, 15, 11, 10, 5, 0]),
+            "expected_data": (
+                "data",
+                [-0.818182, -0.272727, 2.0, 4.0, 5.6, 6.0, 4.0, 1.0],
+            ),
+            "grid_kwargs": {
+                "coords": {"Z": {"center": "depth", "outer": "depth_bnds"}}
+            },
+            "transform_kwargs": {"method": "linear", "target_data": "dens"},
+        },
         "conservative_depth_depth": {
             "input_coord": ("z", [5, 25, 60]),
[…]
