### 2020-09-10 a0d20e1
Implemented rabernats suggestions
