+++ xgcm/test/test_transform.py
@@ -1,4 +1,9 @@
-"""Tests for low transform logic."""
+"""Tests for transform logic.
+The tests are split into three levels:
+low level: functions operating on numpy(numba) level (located in transform.py)
+mid level: function wrappers that operate on xarray objects (located in transform.py)
+high level: API implementations in the grid object (located in grid.py)
+"""
 
 import pytest
 
@@ -13,7 +18,366 @@ from ..transform import (
 
 from xgcm.grid import Grid, Axis
 
-"""Tests for low level routines."""
+
+"""1D Test datasets for various transformations.
+This part of the module is organized as follows:
+
+1. A nested dictionary with all parameters needed to construct a set of test datasets
+2. Fixtures that use a construct_test_input_data function and all or a subset of the dictionary entries from 1. to returns:
+     (input_dataset, grid_kwargs, target, transform_kwargs, expected_dataarray)
+3. Test functions
+
+"""
+# def cases():
+#     return
+
+cases = {
+    "linear_depth_depth": {
+        "input_coord": ("z", [5, 25, 60]),
+        "input_data": (
+            "data",
+            [0.23246861, 0.45175654, 0.58320681],
+        ),  # random numbers
+        "target_coord": ("z", [0, 7, 30, 60, 70]),
+        "target_data": ("z", [0, 7, 30, 60, 70]),
+        "expected_coord": ("z", [0, 7, 30, 60, 70]),
+        "expected_data": (
+            "data",
+            np.interp(
+                [0, 7, 30, 60, 70],
+                [5, 25, 60],
+                [0.23246861, 0.45175654, 0.58320681],
+            ),
+        ),  # same input as in `input_data`
+        "grid_kwargs": {"coords": {"Z": {"center": "z"}}},
+        "transform_kwargs": {"mask_edges": False, "method": "linear"},
+    },
+    "linear_depth_depth_masked_outcrops": {
+        "input_coord": ("z", [5, 25, 60]),
+        "input_data": (
+            "data",
+            [0.23246861, 0.45175654, 0.58320681],
+        ),  # random numbers
+        "target_coord": ("z", [0, 7, 30, 60, 70]),
+        "target_data": ("z", [0, 7, 30, 60, 70]),
+        "expected_coord": ("z", [0, 7, 30, 60, 70]),
+        "expected_data": (
+            "data",
+            np.interp(
+                [0, 7, 30, 60, 70],
+                [5, 25, 60],
+                [0.23246861, 0.45175654, 0.58320681],
+            ),
+        ),  # same input as in `input_data`
+        "expected_data_mask_index": [0, -1],
+        "expected_data_mask_value": np.nan,
+        "grid_kwargs": {"coords": {"Z": {"center": "z"}}},
+        "transform_kwargs": {"mask_edges": True, "method": "linear"},
+    },
+    "linear_depth_depth_renamed": {
+        "input_coord": ("test", [5, 25, 60]),
+        "input_data": (
+            "data",
+            [0.23246861, 0.45175654, 0.58320681],
+        ),  # random numbers
+        "target_coord": ("something", [0, 7, 30, 60, 70]),
+        "target_data": ("something", [0, 7, 30, 60, 70]),
+        "expected_coord": ("something", [0, 7, 30, 60, 70]),
+        "expected_data": (
+            "data",
+            np.interp(
+                [0, 7, 30, 60, 70],
+                [5, 25, 60],
+                [0.23246861, 0.45175654, 0.58320681],
+            ),
+        ),  # same input as in `input_data`
+        "grid_kwargs": {"coords": {"Z": {"center": "test"}}},
[…]
