### 2020-09-10 a7a2177
Added test for autonaming logic, failing currently
