### 2020-09-10 f9b5be1
Fixed auto naming feature
