+++ ci/environment-upstream-dev.yml
@@ -11,4 +11,5 @@ dependencies:
     - docrep<=0.2.7
     - codecov
     - pytest-cov
+    - pytest-xdist
     - git+https://github.com/pydata/xarray.git
