### 2020-09-11 xgcm/xgcm#231 @vrx-
xgcm authorship
This comment is part of the JOSS review at openjournals/joss-reviews#2631

As part of the review process, the [guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#authorship) prompt me to confirm that the full list of paper authors seems appropriate and complete, and to raise an issue if it’s not obviously complete/correct. For example,

* a couple of authors seem to have commits of a single typo fix (geosciz ) and yaml file for release (andersy005) which don't seem to be of obvious relevance for the development of the package
* there are other contributors, with both commits and/or also substantial input in raised issues, who are not part of the author list (raphaeldussin, dcherian, apatlpo)

I understand that contributions may also occur behind the scenes as substantial intellectual input, but as they are not obvious for the external reviewer I'm raising the issue to follow the JOSS guidelines so everything can be out in the open.
