### 2020-09-11 xgcm/xgcm#231 @rabernat
Thanks @vrx-. Authorship was discussed internally in https://github.com/xgcm/xgcm/pull/181#issue-375778797.

The JOSS paper draft predates the contributions of @dcherian and @apatlpo, who have started contributing only in the past few months. Would it be okay to add them as authors at this stage? I think that would be great.

My initial approach was to offer authorship to everyone who has contributed, even if the contribution was small in terms of commits / lines of code. That appears to be acceptable based on the JOSS Guidelines:

> Ultimately, the authors themselves are responsible for deciding which contributions are sufficient for co-authorship, although JOSS policy is that purely financial contributions are not considered sufficient.
