### 2020-09-11 7dda7c0
Create pull_request_template.md (#230)
Copied from xarray
