### 2020-09-11 c5c5fa0
Update .github/ISSUE_TEMPLATE/bug_report.md
Co-authored-by: Ryan Abernathey <ryan.abernathey@gmail.com>
