### 2020-09-11 xgcm/xgcm#231 @vrx-
There isn't anything in the guidelines that prevent adding authors at this stage, so as a reviewer I would encourage to add @dcherian and @apatlpo if they are willing.
