### 2020-09-14 xgcm/xgcm#231 @kthyng
@rabernat and @jbusecke Yes you can add co-authors and should if appropriate, and yes you can choose who can be a co-author. Authorship for JOSS should be determined in a similar way to a traditional publication where all authors have contributed in some notable manner [and](https://joss.readthedocs.io/en/latest/submitting.html#authorship)

> co-authors agree to be accountable for all aspects of the work
