### 2020-09-14 xgcm/xgcm#231 @jbusecke
I will open a PR to add both @dcherian and @apatlpo if they agree.

If you could both send me a confirmation and your affiliation details (if available with ORCID like [here](https://raw.githubusercontent.com/xgcm/xgcm/master/paper.md)), that would be awesome.
