### 2020-09-14 xgcm/xgcm#231 @dcherian
I am happy and grateful to be a coauthor on this paper

```
  - name: Deepak Cherian
    orcid: 0000-0002-6861-8734
    affiliation: 3
```

Also NCAR's affiliation thing is usually "National Center for Atmospheric Research, Boulder, USA"
