### 2020-09-14 8dd14f1
Additional JOSS authors
Adding @dcherian and @apatlpo (pending) to author list for JOSS paper
