### 2020-09-14 24af94a
Add versioneer and setup.py to codecov ignore
