### 2020-09-14 5213171
Consolidate codecov settings
