### 2020-09-14 4b66fea
Update .github/workflows/ci.yaml
Co-authored-by: Anderson Banihirwe <axbanihirwe@ualr.edu>
