+++ .github/workflows/ci.yaml
@@ -81,4 +81,4 @@ jobs:
     - name: Run Tests
       shell: bash -l {0}
       run: |
-        pytest
+        pytest -n auto
