### 2020-09-14 2fc4038
Added failing case to tests
