### 2020-09-14 caba841
Fix for edge case failing
