### 2020-09-15 xgcm/xgcm#231 @jbusecke
Just pinging @apatlpo again. If you have a few minutes and could send me your approval and affiliation, we can get #232 merged.
