+++ xgcm/test/test_transform.py
@@ -6,7 +6,7 @@ high level: API implementations in the grid object (located in grid.py)
 """
 
 import pytest
-
+import dask
 import numpy as np
 import xarray as xr
 from ..transform import (
@@ -23,18 +23,16 @@ from xgcm.grid import Grid, Axis
 This part of the module is organized as follows:
 
 1. A nested dictionary with all parameters needed to construct a set of test datasets
-2. Fixtures that use a construct_test_input_data function and all or a subset of the dictionary entries from 1. to returns:
-     (input_dataset, grid_kwargs, target, transform_kwargs, expected_dataarray)
+2. Fixtures that use a construct_test_source_data function and all or a subset of the dictionary entries from 1. to returns:
+     (source_dataset, grid_kwargs, target, transform_kwargs, expected_dataarray)
 3. Test functions
 
 """
-# def cases():
-#     return
 
 cases = {
     "linear_depth_depth": {
-        "input_coord": ("z", [5, 25, 60]),
-        "input_data": (
+        "source_coord": ("z", [5, 25, 60]),
+        "source_data": (
             "data",
             [0.23246861, 0.45175654, 0.58320681],
         ),  # random numbers
@@ -48,13 +46,13 @@ cases = {
                 [5, 25, 60],
                 [0.23246861, 0.45175654, 0.58320681],
             ),
-        ),  # same input as in `input_data`
+        ),  # same source as in `source_data`
         "grid_kwargs": {"coords": {"Z": {"center": "z"}}},
         "transform_kwargs": {"mask_edges": False, "method": "linear"},
     },
     "linear_depth_depth_masked_outcrops": {
-        "input_coord": ("z", [5, 25, 60]),
-        "input_data": (
+        "source_coord": ("z", [5, 25, 60]),
+        "source_data": (
             "data",
             [0.23246861, 0.45175654, 0.58320681],
         ),  # random numbers
@@ -68,15 +66,15 @@ cases = {
                 [5, 25, 60],
                 [0.23246861, 0.45175654, 0.58320681],
             ),
-        ),  # same input as in `input_data`
+        ),  # same source as in `source_data`
         "expected_data_mask_index": [0, -1],
         "expected_data_mask_value": np.nan,
         "grid_kwargs": {"coords": {"Z": {"center": "z"}}},
         "transform_kwargs": {"mask_edges": True, "method": "linear"},
     },
     "linear_depth_depth_renamed": {
-        "input_coord": ("test", [5, 25, 60]),
-        "input_data": (
+        "source_coord": ("test", [5, 25, 60]),
+        "source_data": (
             "data",
             [0.23246861, 0.45175654, 0.58320681],
         ),  # random numbers
@@ -90,17 +88,32 @@ cases = {
                 [5, 25, 60],
                 [0.23246861, 0.45175654, 0.58320681],
             ),
-        ),  # same input as in `input_data`
+        ),  # same source as in `source_data`
         "grid_kwargs": {"coords": {"Z": {"center": "test"}}},
         "transform_kwargs": {"mask_edges": False, "method": "linear"},
     },
     # example of interpolating onto a tracer that increases with depth
     # but with inverted target
     "linear_depth_dens": {
-        "input_coord": ("depth", [5, 25, 60, 80, 100, 120]),
-        "input_data": ("data", [1, 4, 6, 2, 0, -3]),
[…]
