+++ xgcm/test/test_transform.py
@@ -721,4 +721,4 @@ def test_grid_transform_multidim(request, client, multidim_cases):
 
 # TODO:
 # - What happens when target_data and data are on difference coords? Should we interpolate onto the same.
-# - performance test?
+# - performance test? It would be nice to have these right after so we can track the performance as this feature evolves
