### 2020-09-15 6b82ebc
Added distributed to upstream ci env
