### 2020-09-15 27403ba
Added test for chunking along the transform dimension
