### 2020-09-15 b5e63f8
Implement Ryans suggestions
