### 2020-09-15 e9093ab
Added flip to linear core algo + tests
