+++ xgcm/transform.py
@@ -16,6 +16,11 @@ from numba import jit, guvectorize, float32, float64, boolean
     nopython=True,
 )
 def _interp_1d_linear(phi, theta, target_theta_levels, mask_edges, output):
+    # if last theta value is smaller than first, assume the profile is monotonically decreasing and flip
+    if theta[-1] < theta[0]:
+        theta = np.flip(theta)
+        phi = np.flip(phi)
+
     output[:] = np.interp(target_theta_levels, theta, phi)
 
     if mask_edges:
