### 2020-09-15 4c28e58
Add interpolation step to grid method
