+++ xgcm/test/test_transform.py
@@ -224,12 +224,12 @@ cases = {
             [0, 10, 30, 70, 90, 110, 170],
         ),
         "source_additional_data": ("dens", [1, 5, 10, 20, 24, 35, 37]),
-        "target_coord": ("dens", [0, 5, 36]),
-        "target_data": ("dens", [0, 5, 36]),
-        "expected_coord": ("dens", [2.5, 20.5]),
+        "target_coord": ("dens", [0, 5, 38]),
+        "target_data": ("dens", [0, 5, 38]),
+        "expected_coord": ("dens", [2.5, 21.5]),
         "expected_data": (
             "data",
-            [1, 10.5],
+            [1, 9],
         ),
         "grid_kwargs": {"coords": {"Z": {"center": "depth", "outer": "depth_bnds"}}},
         "transform_kwargs": {
@@ -246,13 +246,13 @@ cases = {
             "depth_bnds",
             [0, 10, 30, 70, 90, 110, 170],
         ),
-        "source_additional_data": ("temp", [37, 35, 24, 20, 10, 5, 1]),
+        "source_additional_data": ("temp", [30, 25, 20, 15, 10, 5, 0]),
         "target_coord": ("temp", [0, 5, 36]),
         "target_data": ("temp", [0, 5, 36]),
         "expected_coord": ("temp", [2.5, 20.5]),
         "expected_data": (
             "data",
-            [1, 10.5],
+            [1, 9],
         ),
         "grid_kwargs": {"coords": {"Z": {"center": "depth", "outer": "depth_bnds"}}},
         "transform_kwargs": {
@@ -260,27 +260,25 @@ cases = {
             "target_data": "temp",
         },
     },
-    # This should error out I think. I think we need to interpolate the additional
-    # dens data on the cell faces. Somehow this does compute though and returns 3 values?
-    # seems like a bug.
-    "conservative_depth_dens": {
+    # Trying to make sure that these results are the same as `conservative_depth_temp_on_bounds`
+    "conservative_depth_temp": {
         "source_coord": ("depth", [5, 25, 60, 80, 100, 120]),
         "source_bounds_coord": ("depth_bnds", [0, 10, 30, 70, 90, 110, 170]),
-        "source_data": ("data", [1, 4, 6, 2, 0, -3]),
+        "source_data": ("data", [-3, 0, 2, 6, 4, 1]),
         "source_additional_data_coord": ("depth", [5, 25, 60, 80, 100, 120]),
-        "source_additional_data": ("dens", [1, 5, 10, 20, 24, 35]),
-        "target_coord": ("dens", [0, 5, 36]),
-        "target_data": ("dens", [0, 5, 36]),
-        "expected_coord": ("dens", [2.5, 7.5]),
+        "source_additional_data": ("temp", [27.5, 22.5, 17.5, 12.5, 7.5, 2.5]),
+        "target_coord": ("temp", [0, 5, 36]),
+        "target_data": ("temp", [0, 5, 36]),
+        "expected_coord": ("temp", [2.5, 20.5]),
         "expected_data": (
             "data",
-            [5.0, 5.0],
+            [1, 9],
         ),
         "grid_kwargs": {"coords": {"Z": {"center": "depth", "outer": "depth_bnds"}}},
-        "error": True,  # this currently fails but shouldnt
+        "error": True,  # this will still fail in the mid level tests but should succeed in the high level version (due to implementing interp on the high level method)
         "transform_kwargs": {
             "method": "conservative",
-            "target_data": "dens",
+            "target_data": "temp",
         },
     },
     # Check for bugfix. Tranform onto column with two or more equal values.
@@ -290,12 +288,12 @@ cases = {
[…]
