### 2020-09-15 ec6ae26
Clean up and trimming input_handling
