### 2020-09-15 983ed58
Bugfix for input_handling
