### 2020-09-15 96221b9
Find oldest numba version that works
