+++ ci/environment-py3.7.yml
@@ -4,7 +4,7 @@ dependencies:
   - xarray
   - dask
   - numpy
-  - numba
+  - numba==0.48
   - pytest
   - future
   - pip
