+++ ci/environment-py3.8.yml
@@ -4,7 +4,7 @@ dependencies:
   - xarray
   - dask
   - numpy
-  - numba
+  - numba==0.49
   - pytest
   - future
   - pip
