+++ ci/environment-py3.6.yml
@@ -4,7 +4,7 @@ dependencies:
   - xarray
   - dask
   - numpy
-  - numba==0.50
+  - numba == 0.49
   - pytest
   - future
   - pip
