### 2020-09-15 0e3b221
Add numba >0.49 as dependency
