### 2020-09-15 10a1141
Parallelize pytest (#226)
* Parallelize pytest

* Update environment-py3.6.yml

* Update environment-py3.7.yml

* Update environment-py3.8.yml

* Update environment-upstream-dev.yml

* Update .github/workflows/ci.yaml

Co-authored-by: Anderson Banihirwe <axbanihirwe@ualr.edu>

Co-authored-by: Anderson Banihirwe <axbanihirwe@ualr.edu>
