+++ ci/environment-py3.7.yml
@@ -12,4 +12,5 @@ dependencies:
     - docrep<=0.2.7
     - codecov
     - pytest-cov
+    - pytest-xdist
 
