+++ .github/workflows/ci.yaml
@@ -43,7 +43,7 @@ jobs:
       - name: Run Tests
         shell: bash -l {0}
         run: |
-          pytest --cov=./ --cov-report=xml
+          pytest -n auto --cov=./ --cov-report=xml
       - name: Upload code coverage to Codecov
         uses: codecov/codecov-action@v1
         with:
@@ -81,4 +81,4 @@ jobs:
     - name: Run Tests
       shell: bash -l {0}
       run: |
-        pytest
+        pytest -n auto
