### 2020-09-16 7b70e53
Making numba optional
