+++ ci/environment-py3.6.yml
@@ -4,7 +4,7 @@ dependencies:
   - xarray
   - dask
   - numpy
-  - numba
+  # no numba to test optional dependency
   - pytest
   - future
   - pip
