+++ .github/workflows/ci.yaml
@@ -17,7 +17,7 @@ jobs:
       fail-fast: false
       matrix:
         os: ["ubuntu-latest"]
-        python-version: ["3.6", "3.7", "3.8"] # 3.9 is not yet released
+        python-version: ["3.6", "3.7", "3.8"]
     steps:
       - uses: actions/checkout@v2
       - name: Cache conda
@@ -27,7 +27,7 @@ jobs:
           CACHE_NUMBER: 0
         with:
           path: ~/conda_pkgs_dir
-          key: ${{ runner.os }}-conda-${{ env.CACHE_NUMBER }}-${{ hashFiles('ci/environment.yml') }}
+          key: ${{ runner.os }}-conda-${{ env.CACHE_NUMBER }}-${{ hashFiles(ci/environment-py${{ matrix.python-version }}.yml) }}
       - uses: goanpeca/setup-miniconda@v1
         with:
           activate-environment: test_env_xgcm # Defined in ci/environment*.yml
@@ -52,6 +52,36 @@ jobs:
           env_vars: OS,PYTHON
           name: codecov-umbrella
           fail_ci_if_error: false
+
+core-deps-dev:
+    name: Build (core-dependencies)
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v2
+    - name: Cache conda
+      uses: actions/cache@v1
+      env:
+        # Increase this value to reset cache if ci/environment-core-deps.yml has not changed
+        CACHE_NUMBER: 0
+      with:
+        path: ~/conda_pkgs_dir
+        key: ${{ runner.os }}-conda-${{ env.CACHE_NUMBER }}-${{ hashFiles('ci/environment-core-deps.yml') }}
+    - uses: goanpeca/setup-miniconda@v1
+      with:
+        activate-environment: test_env_xgcm # Defined in ci/environment-core-deps.yml
+        auto-update-conda: false
+        python-version: 3.8
+        environment-file: ci/environment-core-deps.yml
+        use-only-tar-bz2: true # IMPORTANT: This needs to be set for caching to work properly!
+    - name: Set up conda environment
+      shell: bash -l {0}
+      run: |
+        python -m pip install -e .
+        conda list
+    - name: Run Tests
+      shell: bash -l {0}
+      run: |
+        pytest -n auto
 
   upstream-dev:
     name: Build (upstream-dev)
