### 2020-09-16 4c5f279
Add msg to core-deps env
