+++ ci/environment-core-deps.yml
@@ -1,3 +1,5 @@
+# our test environment for core dependencies only (atm this excludes numba),
+# using the latest stable versions of packages.
 name: test_env_xgcm
 dependencies:
   - python=3.8
