+++ .github/workflows/ci.yaml
@@ -27,7 +27,7 @@ jobs:
           CACHE_NUMBER: 0
         with:
           path: ~/conda_pkgs_dir
-          key: ${{ runner.os }}-conda-${{ env.CACHE_NUMBER }}-${{ hashFiles(ci/environment-py.yml) }}
+          key: ${{ runner.os }}-conda-${{ env.CACHE_NUMBER }}-${{ hashFiles('ci/environment-py${{ matrix.python-version }}.yml') }}
       - uses: goanpeca/setup-miniconda@v1
         with:
           activate-environment: test_env_xgcm # Defined in ci/environment*.yml
