### 2020-09-16 15d2f1e
Add docstring to axis/grid methods
