### 2020-09-16 0cfb3af
Support any monotonic order for target in conservative method + tests
