### 2020-09-16 da0f62d
Switch default for mask_edges
