### 2020-09-16 5045944
Fix test settings for new masking default
