### 2020-09-16 50ba361
Check numba version explicitly
