### 2020-09-16 51b2051
Overhauled docstring
