### 2020-09-16 db6bd27
Fix error message for multidim target
