### 2020-09-18 065acfb
Fix naming of output arrays + tests
