### 2020-09-18 1b71d1f
Add entry to whatsnew
