### 2020-09-18 ed98cfd
Some nits on the whatsnew
