### 2020-09-18 c286254
handle missing values in theta
