### 2020-09-18 a6242a4
Fix handling of da with no name
