### 2020-09-18 51aa783
fix bug and add tests
