+++ xgcm/transform.py
@@ -78,13 +78,16 @@ def _interp_1d_conservative(phi, theta_1, theta_2, theta_hat_1, theta_hat_2, out
 
         # handle missing values
         if np.isnan(theta_1[i]) and np.isnan(theta_2[i]):
-            break
+            print("Both NaNs")
+            continue
+        # in the next two cases, we are effectively applying a boundary condition
+        # by assuming that theta is homogenous over the cell
         elif np.isnan(theta_1[i]):
             theta_min = theta_max = theta_2[i]
         elif np.isnan(theta_2[i]):
             theta_min = theta_max = theta_1[i]
         # handle non-monotonic stratification
-        if theta_1[i] < theta_2[i]:
+        elif theta_1[i] < theta_2[i]:
             theta_min = theta_1[i]
             theta_max = theta_2[i]
         else:
