+++ xgcm/transform.py
@@ -78,7 +78,6 @@ def _interp_1d_conservative(phi, theta_1, theta_2, theta_hat_1, theta_hat_2, out
 
         # handle missing values
         if np.isnan(theta_1[i]) and np.isnan(theta_2[i]):
-            print("Both NaNs")
             continue
         # in the next two cases, we are effectively applying a boundary condition
         # by assuming that theta is homogenous over the cell
