+++ doc/index.rst
@@ -45,6 +45,7 @@ Contents
    grids
    grid_topology
    grid_metrics
+   transform
    example_mitgcm
    example_eccov4
    example_MOM6
