### 2020-09-21 41d0f9b
Added alignment checking + Test; issues with manually broadcasted target_data
