### 2020-09-22 xgcm/xgcm#231 @apatlpo
I failed to see all the pings, so sorry about this.

thanks a lot for this kind offer which I won't turn down if you think it is deserved.
please let me know if I can help anyhow with the paper

I apologize in advance for the super long affiliation (crazy french thing):

name: Aurélien  Ponte
ORCID: 0000-0002-0252-6028
Affiliation: Univ. Brest, CNRS, IRD, Ifremer, Laboratoire d’Océanographie Physique et Spatiale (LOPS), IUEM, Brest, France
