### 2020-09-22 xgcm/xgcm#231 @vrx-
> @vrx- I have merged #232. Do you think we can close this issue?

yes!
