### 2020-09-23 723414f
Update doc/whats-new.rst
Co-authored-by: Deepak Cherian <dcherian@users.noreply.github.com>
