### 2020-09-23 a608160
narrower docstring
