+++ xgcm/grid.py
@@ -793,12 +793,12 @@ class Axis:
         transformation:
 
         - 'linear': Values are linear interpolated between 1D columns
-        along `axis` of `da` and `target_data`. This methodrequires 
-        `target_data` to increase/decrease monotonically. `target` 
-        values are interpreted as new cell centers in this case. By 
+        along `axis` of `da` and `target_data`. This methodrequires
+        `target_data` to increase/decrease monotonically. `target`
+        values are interpreted as new cell centers in this case. By
         default this method will return nan for values in `target` that
-        are outside of the range of `target_data`, setting 
-        `mask_edges=False` results in the default np.interp behavior of 
+        are outside of the range of `target_data`, setting
+        `mask_edges=False` results in the default np.interp behavior of
         repeated values.
 
         - 'conservative': Values are transformed while conserving the
@@ -807,7 +807,7 @@ class Axis:
         only work with extensive quantities (like heat, mass, transport)
         but not with intensive quantities (like temperature, density,
         velocity). N given `target` values are interpreted as cell-bounds
-        and the returned array will have N-1 elements along the newly 
+        and the returned array will have N-1 elements along the newly
         created coordinate, with coordinate values that are interpolated
         between `target` values.
 
@@ -1750,12 +1750,12 @@ class Grid:
         transformation:
 
         - 'linear': Values are linear interpolated between 1D columns
-        along `axis` of `da` and `target_data`. This methodrequires 
-        `target_data` to increase/decrease monotonically. `target` 
-        values are interpreted as new cell centers in this case. By 
+        along `axis` of `da` and `target_data`. This methodrequires
+        `target_data` to increase/decrease monotonically. `target`
+        values are interpreted as new cell centers in this case. By
         default this method will return nan for values in `target` that
-        are outside of the range of `target_data`, setting 
-        `mask_edges=False` results in the default np.interp behavior of 
+        are outside of the range of `target_data`, setting
+        `mask_edges=False` results in the default np.interp behavior of
         repeated values.
 
         - 'conservative': Values are transformed while conserving the
@@ -1764,7 +1764,7 @@ class Grid:
         only work with extensive quantities (like heat, mass, transport)
         but not with intensive quantities (like temperature, density,
         velocity). N given `target` values are interpreted as cell-bounds
-        and the returned array will have N-1 elements along the newly 
+        and the returned array will have N-1 elements along the newly
         created coordinate, with coordinate values that are interpolated
         between `target` values.
 
