### 2020-09-23 2272ee7
Update xgcm/grid.py
Co-authored-by: Ryan Abernathey <ryan.abernathey@gmail.com>
