+++ xgcm/grid.py
@@ -1750,7 +1750,7 @@ class Grid:
         transformation:
 
         - 'linear': Values are linear interpolated between 1D columns
-        along `axis` of `da` and `target_data`. This methodrequires
+          along `axis` of `da` and `target_data`. This method requires
         `target_data` to increase/decrease monotonically. `target`
         values are interpreted as new cell centers in this case. By
         default this method will return nan for values in `target` that
