### 2020-09-23 763dcfc
stray space in grid docstring
