### 2020-09-23 8082b38
adjusted grid/axis docstring again
