### 2020-09-23 4e90d42
Update environment-py3.6.yml
