### 2020-09-23 69557ee
Update xgcm/grid.py
Co-authored-by: Ryan Abernathey <ryan.abernathey@gmail.com>
