### 2020-09-23 b5578ee
Update grid.py
