### 2020-09-24 e0eae1e
Optimized vertical grid routines (#205)
* add vertical grid routines

* add numba to ci

* more ci fix

* add numba to dev env

* Fix ci back

* Remove old vert files. Add axis method

* Added grid method

* Added modules back

* bugfix for the datasets fixture

* Add numba to doc env

* move input handling to decorator

* Added first hacky test for conservative interpolation

* Hotfix, grid level conservative still not working

* Refactored test dataset creation

* Wrapped cases dict in function to avoid modification

* Add conservative examples depth-depth

* Conservative method tests with w
[…]
