+++ xgcm/test/test_transform.py
@@ -1,0 +1,970 @@
+"""Tests for transform logic.
+The tests are split into three levels:
+low level: functions operating on numpy(numba) level (located in transform.py)
+mid level: function wrappers that operate on xarray objects (located in transform.py)
+high level: API implementations in the grid object (located in grid.py)
+"""
+
+# TODO:
+# - performance test? It would be nice to have these right after so we can track the performance as this feature evolves
+
+import pytest
+import dask
+import numpy as np
+import xarray as xr
+from xgcm.grid import Grid, Axis
+
+try:
+    import numba
+    from ..transform import (
+        interp_1d_linear,
+        interp_1d_conservative,
+        linear_interpolation,
+        conservative_interpolation,
+    )
+except ImportError:
+    numba = None
+
+
+"""1D Test datasets for various transformations.
+This part of the module is organized as follows:
+
+1. A nested dictionary with all parameters needed to construct a set of test datasets
+2. Fixtures that use a construct_test_source_data function and all or a subset of the dictionary entries from 1. to returns:
+     (source_dataset, grid_kwargs, target, transform_kwargs, expected_dataarray)
+3. Test functions
+
+"""
+
+cases = {
+    "linear_depth_depth": {
+        "source_coord": ("z", [5, 25, 60]),
+        "source_data": (
+            "data",
+            [0.23246861, 0.45175654, 0.58320681],
+        ),  # random numbers
+        "target_coord": ("z", [0, 7, 30, 60, 70]),
+        "target_data": ("z", [0, 7, 30, 60, 70]),
+        "expected_coord": ("z", [0, 7, 30, 60, 70]),
+        "expected_data": (
+            "data",
+            np.interp(
+                [0, 7, 30, 60, 70],
+                [5, 25, 60],
+                [0.23246861, 0.45175654, 0.58320681],
+            ),
+        ),  # same source as in `source_data`
+        "expected_data_mask_index": [0, -1],
+        "expected_data_mask_value": np.nan,
+        "grid_kwargs": {"coords": {"Z": {"center": "z"}}},
+        "transform_kwargs": {"mask_edges": True, "method": "linear"},
+    },
+    "linear_depth_depth_nomask": {
+        "source_coord": ("z", [5, 25, 60]),
+        "source_data": (
+            "data",
+            [0.23246861, 0.45175654, 0.58320681],
+        ),  # random numbers
+        "target_coord": ("z", [0, 7, 30, 60, 70]),
+        "target_data": ("z", [0, 7, 30, 60, 70]),
+        "expected_coord": ("z", [0, 7, 30, 60, 70]),
+        "expected_data": (
+            "data",
+            np.interp(
+                [0, 7, 30, 60, 70],
+                [5, 25, 60],
+                [0.23246861, 0.45175654, 0.58320681],
+            ),
+        ),  # same source as in `source_data`
+        "grid_kwargs": {"coords": {"Z": {"center": "z"}}},
+        "transform_kwargs": {"mask_edges": False, "method": "linear"},
+    },
+    "linear_depth_depth_renamed": {
+        "source_coord": ("test", [5, 25, 60]),
+        "source_data": (
+            "data",
+            [0.23246861, 0.45175654, 0.58320681],
+        ),  # random numbers
+        "target_coord": ("something", [0, 7, 30, 60, 70]),
+        "target_data": ("something", [0, 7, 30, 60, 70]),
[…]
