+++ ci/environment-core-deps.yml
@@ -1,10 +1,11 @@
+# our test environment for core dependencies only (atm this excludes numba),
+# using the latest stable versions of packages.
 name: test_env_xgcm
 dependencies:
-  - python=3.6
+  - python=3.8
   - xarray
   - dask
   - numpy
-  - numba
   - pytest
   - future
   - pip
@@ -14,4 +15,3 @@ dependencies:
     - codecov
     - pytest-cov
     - pytest-xdist
-
