+++ ci/environment-upstream-dev.yml
@@ -1,7 +1,6 @@
 name: test_env_xgcm
 dependencies:
   - python
-  - dask
   - numpy
   - pytest
   - future
@@ -13,3 +12,7 @@ dependencies:
     - pytest-cov
     - pytest-xdist
     - git+https://github.com/pydata/xarray.git
+    - git+https://github.com/dask/dask.git
+    - git+https://github.com/dask/distributed.git
+    - toolz
+    - git+https://github.com/numba/numba.git
