### 2020-09-28 xgcm/xgcm#243 @koldunovn
example_notebooks [BUG]
**What happened**:

There is a very old folder `example_notebooks` that contains only one notebook that nostalgically imports `xray` instead of `xarray` :)

**What you expected to happen**:

Would you consider to just revomove this folder entirelly, to avid confusion for people like me, who would like to run notebooks with examples the first thing after installation? Or maybe better link the notebooks from the documentation to this folder. At least to me `doc` folder is not the most obvious place to look for example notebooks.
