### 2020-09-28 xgcm/xgcm#243 @jbusecke
Oh wow that must be ancient (It has not been touched for 5 years 😱).

> Would you consider to just revomove this folder entirelly

Yup that should be removed for sure. Any objections @rabernat?
