### 2020-09-28 ed2c6a0
Update whats-new.rst
