### 2020-09-28 38ccaa0
Update whats-new.rst
