### 2020-09-28 xgcm/xgcm#246 @malmans2
I'm pretty sure it's new matlotlib default shading (`auto`): https://github.com/SciTools/cartopy/issues/1638
You can still specify `flat`, but I think it will be deprecated soon.
If I'm understating it right, the problem should go away if you use the edge coordinates instead (i.e., use XG and YG if you are plotting variables on the C grid).
