### 2020-09-28 xgcm/xgcm#249 @jbusecke
Thanks again for pointing out this issue with the documentation.

This could be related to the transition of the pangeo cloud deployments. Ill have a look soon (probably wont have time for it today though).
