### 2020-09-28 2bf05cf
Added test to verify global vs method input for boundary and fill_value; failing
