### 2020-09-28 #2631 @koldunovn
There are just a couple things left:

Issues: https://github.com/xgcm/xgcm/issues/243 https://github.com/xgcm/xgcm/issues/246 https://github.com/xgcm/xgcm/issues/249

Comments to the text:

**Statement of Need**
I think the statement "Xgcm supports orthogonal curvilinear grids, including complex, multiply-connected grid topologies such as the cubed-sphere." should be moved to the beggining of the "Statement of Need", espetially since authors suggest that package will be useful for "plasma physics or astrophysical modeling" where other grid types are quite common. This will more clearly position the package.

**State of the field**
To my knowlege there are no packages with similar functionality in the python world. But there are similar efforts, at least with respect of trying to get boundaries right, and properly integrate values, for example,  in cdo (https://code.mpimet.mpg.de/projects/cdo/wiki/Cdo#Documentation). NCL (https://www.ncl.ucar.edu/index.shtml), to my knowlege, also understands grid staggering, and, I gues, this will be also true for its sucecsor https://github.com/NCAR/geocat-comp .
