### 2020-09-28 3247473
Update xgcm/test/test_metrics_ops_single_axis.py
Co-authored-by: Deepak Cherian <dcherian@users.noreply.github.com>
