### 2020-09-29 #2631 @rabernat
Hi all. I'm very grateful for the work of the editors and reviewers so far on this paper. Thanks so much for your time!

The authors have conferred and have decided that we would like to suspend (if possible) or withdraw this paper. We have concluded that publishing a paper about xgcm with an all-male author list is not a good idea. I wish we had realized this before submitting, but the recent NumPy paper fiasco helped open our eyes to the potential harm this can cause.

We have several concrete plans to increase the gender diversity of our contributors and hopefully add some more authors. But this will take time, possibly a year.

I definitely do **not** want to waste the time the reviewers have already spent reviewing this paper. Ideally what we would like to do is to continue this review process, to the point where we have all the boxes checked, but pause going forward to the publication stage. At a future date, hopefully we can pick things up from where we left off.

@kthyng--is this compatible with JOSS policies? If not, what course of action would you recommend?

Sorry for introducing this fly in the ointment to an otherwise very smooth review process. We are trying, imperfectly, to do the right thing.
