### 2020-09-30 #2631 @kthyng
@rabernat I understand your concern. I think the best course of action is to heartily thank the reviewers for their time. xgcm should be reviewed at the time you want to publish because all of the software at the time of publishing should be reviewed, not just what is in the repo today. The future authors you hope to recruit will make changes that should be included in the review.

So, my suggestion is to withdraw the paper now with the intention of resubmitting at a later date. Is this what you would like to do?
