### 2020-10-01 9ca1c73
First implemntation of `reverse` for cumsum; left/right tests fail
