### 2020-10-06 19d0fba
Replaced manual computation with xr.weighted; all tests pass.
