### 2020-10-06 d1eb097
fix; test comparison bumped down from equals to allclose.
