### 2020-10-06 #2631 @rabernat
Sorry for the delay, and thanks for everyone's patience.

Yes, we have decided to withdraw the paper. I'm **extremely grateful** for the time you have spent on this, and I **apologize deeply and sincerely** for not seeing this glaring problem _before_ submitting our paper. The reviewers' time was definitely not wasted from my point of view, as the comments have already led to considerably improvements of the package.
